projeto finalizado

+++ src/controllers/QuestionController.js
@@ -2,6 +2,7 @@ const Database = require('../db/config')
 
 module.exports = {
   async index(req, res) {
+    const db = await Database()
     const roomId = req.params.room
     const questionId = req.params.question
     const action = req.params.action
@@ -9,7 +10,7 @@ module.exports = {
 
     /* Verificar se a senha está correta */
     const verifyRoom = await db.get(`SELECT * FROM rooms WHERE id = ${roomId}`)
-    if (verifyRoom.pass == password) {
+    if (verifyRoom && verifyRoom.pass == password) {
       if (action == 'delete') {
         await db.run(`DELETE FROM questions WHERE id = ${questionId}`)
       } else if (action == 'check') {
@@ -27,14 +28,14 @@ module.exports = {
     const roomId = req.params.room
 
     await db.run(`INSERT INTO questions(
-        title,
-        room,
-        read
-    )VALUES(
-        "${question}",
-        ${roomId},
-        0
-    )`)
+            title,
+            room,
+            read
+        )VALUES(
+            "${question}",
+            ${roomId},
+            0
+        )`)
 
     res.redirect(`/room/${roomId}`)
   }
